backend/requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
Django==4.2.16
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
whitenoise==6.7.0
